Add second test user (id: 12345)

+++ server/lib/tests/setup.ts
@@ -2,7 +2,7 @@ import { MySQL } from 'lib/MySQL';
 
 export async function setupTests() {
   const db = new MySQL();
-  await db.query('DELETE FROM users WHERE userId = 1234');
-  await db.query('INSERT INTO users (userId) VALUES (1234)');
+  await db.query('DELETE FROM users WHERE userId = 1234 OR userId = 12345');
+  await db.query('INSERT INTO users (userId) VALUES (1234), (12345)');
   db.release();
 }
